bento/requirements.txt
numpy==1.19.2
scikit-learn==0.23.2
tensorflow==2.3.0
pandas==1.1.3
Pillow==7.2.0